Fix build status dialog and texts
