+++ components/settings/impl/src/main/java/com/flipperdevices/settings/impl/composable/category/ExperimentalCategory.kt
@@ -8,6 +8,7 @@ import com.flipperdevices.settings.impl.R
 import com.flipperdevices.settings.impl.composable.elements.CategoryElement
 import com.flipperdevices.settings.impl.composable.elements.ClickableElement
 import com.flipperdevices.settings.impl.composable.elements.GrayDivider
+import com.flipperdevices.settings.impl.composable.elements.SwitchableElement
 import com.flipperdevices.settings.impl.viewmodels.ExperimentalViewModel
 import com.flipperdevices.settings.impl.viewmodels.SettingsViewModel
 import tangle.viewmodel.compose.tangleViewModel
@@ -40,6 +41,12 @@ fun ExperimentalCategory(
                 onClick = {
                     experimentalViewModel.onOpenScreenStreaming(navController)
                 }
+            )
+            GrayDivider()
+            SwitchableElement(
+                titleId = R.string.experimental_application_catalog,
+                state = settings.applicationCatalog,
+                onSwitchState = experimentalViewModel::onSwitchApplicationCatalog
             )
         }
     }
