+++ components/settings/impl/src/main/java/com/flipperdevices/settings/impl/viewmodels/ExperimentalViewModel.kt
@@ -1,18 +1,23 @@
 package com.flipperdevices.settings.impl.viewmodels
 
+import androidx.datastore.core.DataStore
+import androidx.lifecycle.viewModelScope
 import androidx.navigation.NavController
 import com.flipperdevices.core.di.provideDelegate
+import com.flipperdevices.core.preference.pb.Settings
 import com.flipperdevices.core.ui.lifecycle.LifecycleViewModel
 import com.flipperdevices.filemanager.api.navigation.FileManagerEntry
 import com.flipperdevices.metric.api.MetricApi
 import com.flipperdevices.metric.api.events.SimpleEvent
 import com.flipperdevices.settings.api.SettingsFeatureEntry
-import tangle.viewmodel.VMInject
 import javax.inject.Provider
+import kotlinx.coroutines.launch
+import tangle.viewmodel.VMInject
 
 class ExperimentalViewModel @VMInject constructor(
     fileManagerEntryProvider: Provider<FileManagerEntry>,
     metricApiProvider: Provider<MetricApi>,
+    private val settingsDataStore: DataStore<Settings>,
     private val settingsFeatureEntry: SettingsFeatureEntry
 ) : LifecycleViewModel() {
     private val fileManagerEntry by fileManagerEntryProvider
@@ -26,5 +31,15 @@ class ExperimentalViewModel @VMInject constructor(
     fun onOpenScreenStreaming(navController: NavController) {
         metricApi.reportSimpleEvent(SimpleEvent.EXPERIMENTAL_OPEN_SCREEN_STREAMING)
         navController.navigate(settingsFeatureEntry.remoteControlStub())
+    }
+
+    fun onSwitchApplicationCatalog(enabled: Boolean) {
+        viewModelScope.launch {
+            settingsDataStore.updateData {
+                it.toBuilder()
+                    .setApplicationCatalog(enabled)
+                    .build()
+            }
+        }
     }
 }
